requirements.txt
# Основные зависимости для проекта trading_predictor

# Работа с данными
pandas>=1.5.0
numpy>=1.24.0

# HTTP запросы
requests>=2.28.0

# Работа с YAML конфигурацией
PyYAML>=6.0

# Jupyter notebook (для разработки)
jupyter>=1.0.0
ipykernel>=6.0.0

# Визуализация (для EDA)
matplotlib>=3.6.0
seaborn>=0.12.0

# Машинное обучение (планируется)
scikit-learn>=1.2.0

# Финансовые данные и анализ (планируется)
yfinance>=0.2.0
ta>=0.10.0

# =============================================================================
# FEATURE ENGINEERING DEPENDENCIES
# =============================================================================

# Статистические вычисления для feature engineering
scipy>=1.9.0

# Технические индикаторы (альтернатива TA-Lib)
pandas-ta>=0.3.14b0

# Работа с Parquet файлами (эффективное хранение)
pyarrow>=12.0.0
fastparquet>=0.8.3

# Тестирование модулей feature engineering
pytest>=7.0.0
pytest-cov>=4.0.0

# =============================================================================
# MODELING DEPENDENCIES
# =============================================================================

# Градиентный бустинг (мультикласс)
xgboost>=1.7.6
lightgbm>=4.3.0
catboost>=1.2

# Сериализация моделей
joblib>=1.3.2